feat(products): refactor product components and routing
