+++ src/components/products/ProductsList.vue
@@ -40,6 +40,7 @@
   </v-container>
 </template>
 
+
 <script setup lang="ts">
 //Router
 import { RouterLink } from "vue-router";
